ActivityPub.Actor (bonfire_umbrella v0.4.0-classic-beta.44)

An ActivityPub Actor type and functions for dealing with actors.

Types

t()

@type t() :: %ActivityPub.Actor{
  ap_id: term(),
  data: term(),
  deactivated: term(),
  id: term(),
  keys: term(),
  local: term(),
  pointer: term(),
  pointer_id: term(),
  username: term()
}

Functions

actor_url(username)

actor?(object)

check_actor_is_active(actor)

deactivate(actor)

delete(actor)

ensure_keys_present(actor)

Checks if an actor struct has a non-nil keys field and generates a PEM if it doesn't.

fetch_by_username(username)

Fetches a remote actor by username in username@domain.tld format

format_remote_actor(object)

format_username(other)

get_cached!(opts)

get_cached(opts)

get_external_followers(actor)

get_followers(actor)

get_followings(actor)

get_or_fetch(username_or_uri)

get_or_fetch_by_ap_id(ap_id, maybe_create \\ true)

get_or_fetch_by_username(username)

Tries to get a local actor by username or tries to fetch it remotely if username is provided in username@domain.tld format.

get_public_key_for_ap_id(ap_id)

Fetches the public key for given actor AP ID.

invalidate_cache(actor)

maybe_create_actor_from_object(actor)

reactivate(actor)

set_cache(actor)

update_actor(actor_id)

@spec update_actor(String.t()) :: {:ok, t()} | {:error, any()}

Updates an existing actor struct by its AP ID.

update_actor(actor_id, data)

update_actor_data_by_ap_id(ap_id, data)