ActivityPub.Federator.APPublisher (bonfire_umbrella v0.9.8-cooperation-beta.24)

Functions

determine_inbox(user, is_public, type, num_recipients)

Determine a user inbox to use based on heuristics. These heuristics are based on an approximation of the sharedInbox rules in the ActivityPub specification.

gather_webfinger_links(actor)

Callback implementation for ActivityPub.Federator.Publisher.gather_webfinger_links/1.

is_representable?(activity)

Callback implementation for ActivityPub.Federator.Publisher.is_representable?/1.

maybe_federate_to_search_index(recipients, activity)

If you put the URL of the shared inbox of an ActivityPub instance in the following env variable, all public content will be pushed there via AP federation for search indexing purposes: PUSH_ALL_PUBLIC_CONTENT_TO_INSTANCE #TODO: move to adapter

publish(actor, activity)

Callback implementation for ActivityPub.Federator.Publisher.publish/2.

publish_one(params)

Publish a single message to a peer. Takes a struct with the following parameters set:

inbox: the inbox to publish to

json: the JSON message body representing the ActivityPub message

actor: the actor which is signing the message

id: the ActivityStreams URI of the message