ActivityPub.Federator.Adapter behaviour (bonfire_umbrella v0.9.8-cooperation-beta.11)

Contract for ActivityPub module adapters

Callbacks

base_url()

@callback base_url() :: String.t()

external_followers_for_activity(t, t)

@callback external_followers_for_activity(List.t(), Map.t()) :: List.t()

Compute and return a subset of followers that should receive a specific activity (optional)

get_actor_by_ap_id(t)

@callback get_actor_by_ap_id(String.t()) :: {:ok, ActivityPub.Actor.t()} | {:error, any()}

get_actor_by_id(t)

@callback get_actor_by_id(String.t()) :: {:ok, ActivityPub.Actor.t()} | {:error, any()}

get_actor_by_username(t)

@callback get_actor_by_username(String.t()) ::
  {:ok, ActivityPub.Actor.t()} | {:error, any()}

Fetch an actor given its preferred username

get_follower_local_ids(t)

@callback get_follower_local_ids(ActivityPub.Actor.t()) :: [String.t()]

get_following_local_ids(t)

@callback get_following_local_ids(ActivityPub.Actor.t()) :: [String.t()]

get_locale()

@callback get_locale() :: String.t()

get_or_create_service_actor()

@callback get_or_create_service_actor() :: ActivityPub.Actor.t() | nil

Creates an internal service actor by username, if missing.

get_redirect_url(arg1)

@callback get_redirect_url(String.t() | Map.t()) :: String.t()

Gets local url of an AP object to redirect in browser. Can take pointer id or an actor username.

handle_activity(t)

@callback handle_activity(ActivityPub.Object.t()) :: :ok | {:ok, any()} | {:error, any()}

Passes data to be handled by the host application

maybe_create_remote_actor(t)

@callback maybe_create_remote_actor(ActivityPub.Actor.t()) :: :ok

maybe_publish_object(t, t)

@callback maybe_publish_object(String.t(), Boolean.t()) :: {:ok, any()} | {:error, any()}

update_local_actor(t, t)

@callback update_local_actor(ActivityPub.Actor.t(), Map.t()) ::
  {:ok, ActivityPub.Actor.t()} | {:error, any()}

update_remote_actor(t)

@callback update_remote_actor(ActivityPub.Object.t()) :: :ok | {:error, any()}

Functions

actor_federating?(actor)

adapter()

base_url()

call_or(fun, args \\ [], fallback \\ nil)

Run function from adapter if defined, otherwise return fallback value

external_followers_for_activity(actor, activity)

get_actor_by_ap_id(id)

get_actor_by_id(id)

get_actor_by_username(username)

get_follower_local_ids(actor)

get_following_local_ids(actor)

get_locale()

get_or_create_service_actor()

get_redirect_url(id_or_username_or_object)

handle_activity(activity)

maybe_create_remote_actor(actor)

maybe_handle_activity(activity)

maybe_publish_object(object, manually_fetching? \\ false)

update_local_actor(actor, params)

update_remote_actor(actor)

update_remote_actor(actor, data)