ActivityPub.Federator.Transformer (bonfire_umbrella v0.9.8-cooperation-beta.24)

This module normalises outgoing data to conform with AS2/AP specs and handles incoming objects and activities

Functions

add_emoji_tags(object)

determine_explicit_mentions(object)

@spec determine_explicit_mentions(map()) :: [any()]

fix_actor(data)

fix_addressing(object)

fix_addressing_list(map, field)

fix_attachments(object)

fix_content_map(object)

fix_context(object)

fix_emoji(object)

fix_explicit_addressing(object, follower_collection)

fix_in_reply_to(object, options \\ [])

fix_object(object, options \\ [])

Modifies an incoming AP object (mastodon format) to our internal format.

fix_quote_url(object, options \\ [])

fix_replies(data, options)

fix_summary(object)

fix_tag(object)

fix_url(object)

format_input(text, binary, options \\ [])

handle_incoming(data)

Handles incoming data, inserts it into the database and triggers side effects if the data is a supported activity type.

maybe_handle_other_activity(data)

maybe_handle_other_object(data)

prepare_outgoing(data)

Translates MN Entity to an AP compatible format

preserve_privacy_of_outgoing(other, target_instance_uri \\ nil)

replies(arg1)

set_replies(object)

Serialized Mastodon-compatible replies collection containing self-replies. Based on Mastodon's ActivityPub::NoteSerializer#replies.

take_emoji_tags(arg1)