ActivityPub.Utils (bonfire_umbrella v0.9.5-classic-beta.1)

Misc functions used for federation

Functions

adapter_fallback()

ap_base_url()

as_local_public()

cachex_fetch(cache, key, fallback, options \\ [])

error_json(conn, error, status \\ 500)

generate_id(type)

generate_object_id()

get_with_cache(get_fun, cache_bucket, key, identifier)

is_ulid?(str)

is_uuid?(str)

Takes a string and returns true if it is a valid UUID (Universally Unique Identifier)

json_with_cache(conn, get_fun, cache_bucket, id)

label_in_message?(label, params)

@spec label_in_message?(String.t(), map()) :: boolean()

make_date()

make_json_ld_header()

maybe_put(map, key, value)

conditionally update a map

maybe_to_atom(str)

ok_unwrap(val, fallback \\ nil)

public?(data)

public?(activity_data, object_data)

Determines if an object or an activity is public.

put_if_present(map, key, value, value_function \\ &{:ok, &1})

repo()

request_ip(conn)

safe_put_in(data, keys, value)

set_repo(repo)

single_ap_id(ap_id)

single_ap_id_or_object(ap_id)

ulid(input)