ActivityPub.Web.ActivityPubController (bonfire_umbrella v0.9.6-classic-beta.2)

Endpoints for serving objects and collections, so the ActivityPub API can be used to read information from the server.

Even though we store the data in AS format, some changes need to be applied to the entity before serving it in the AP REST response. This is done in ActivityPub.Web.ActivityPubView.

actor(conn, map)

ap_route_helper(uuid)

followers(conn, map)

following(conn, map)

inbox(conn, params)

json_object_with_cache(conn \\ nil, id)

noop(conn, params)

object(conn, map)

outbox(conn, map)