ActivityPubWeb.Transmogrifier (bonfire_umbrella v0.9.0-classic-beta.0)

This module normalises outgoing data to conform with AS2/AP specs and handles incoming objects and activities

Functions

add_emoji_tags(object)

allowed_thread_distance?(distance)

Returns true if the distance to target object does not exceed max configured value. Serves to prevent fetching of very long threads, especially useful on smaller instances. Addresses memory leaks on recursive replies fetching. Applies to fetching of both ancestor (reply-to) and child (reply) objects.

determine_explicit_mentions(object)

@spec determine_explicit_mentions(map()) :: [any()]

fix_actor(data)

fix_addressing(object)

fix_addressing_list(map, field)

fix_attachments(object)

fix_content_map(object)

fix_context(object)

fix_emoji(object)

fix_explicit_addressing(object, follower_collection)

fix_in_reply_to(object, options \\ [])

fix_object(object, options \\ [])

Modifies an incoming AP object (mastodon format) to our internal format.

fix_quote_url(object, options \\ [])

fix_summary(object)

fix_tag(object)

fix_url(object)

handle_incoming(data)

Handles incoming data, inserts it into the database and triggers side effects if the data is a supported activity type.

maybe_handle_other(data)

Normalises and inserts an incoming AS2 object. Returns Object.

prepare_object(object)

prepare_outgoing(data)

Translates MN Entity to an AP compatible format

replies(arg1)

set_replies(object)

Serialized Mastodon-compatible replies collection containing self-replies. Based on Mastodon's ActivityPub::NoteSerializer#replies.

take_emoji_tags(arg1)