Arrows (bonfire_umbrella v0.9.5-classic-beta.24)

A handful of (mostly) arrow macros with superpowers.

l <|> r
(macro)

Like ||, except only defaults if the left is nil (i.e. false is valid)

l <~> r
(macro)

Like ||, except with the logic applied by ~>

from_ok(x)

ok(x)

ok_or(x, err)

to_ok(x)

l |> r
(macro)

A more flexible drop-in replacement for the standard elixir pipe operator.

Special features are unlocked when using the ... (ellipsis) on the right hand side:

The right hand side need not be a function, it can be any expression containing the ellipsis.

The ellipsis will be replaced with the result of evaluating the hand side expression.

You may use the ellipsis multiple times and the left hand side will be calculated exactly once.

You can do crazy stuff with the ellipsis, but remember that people have to read it!

l ~> r
(macro)

Like OK.~>