Bonfire.Boundaries (bonfire_umbrella v0.9.0-classic-beta.5)

Functions

acls_from_preset_boundary_names(presets)

boundaries_set(text)

can?(subject, can_verbs?, object_boundary)

list_grants_on(things)

list_grants_on(things, verbs)

eg: list_grants_on(id, [:see, :read])

list_object_acls(object)

load_pointer(item, opts)

load_pointers(items, opts)

Loads binaries according to boundaries (which are assumed to be ULID pointer IDs). Lists which are iterated and return a [sub]list with only permitted pointers.

my_grants_on(users, things)

pointer_permitted?(item, opts)

preset_boundary_from_acl(summary)

preset_boundary_role_from_acl(summary)

preset_boundary_tuple_from_acl(acl)

preset_name(boundaries)

set_boundaries(creator, object, opts)

take_care_of!(things, user)

Assigns the user as the caretaker of the given object or objects, replacing the existing caretaker, if any.

user_default_boundaries()