Bonfire.Boundaries.Acls (bonfire_umbrella v0.4.0-classic-beta.35)

acls represent fully populated access control rules that can be reused

Functions

acls()

built_in_ids()

cast(changeset, creator, opts)

changeset(atom, attrs, opts)

create(attrs \\ %{}, opts)

delete(acl, opts)

Fully delete the ACL, including permissions/grants and controlled information. This will affect all objects previously shared with this ACL.

edit(acl, user, params)

find_caretaker_stereotypes(caretaker, stereotypes)

for_dropdown(opts)

get!(slug)

get(slug)

get_for_caretaker(id, caretaker, opts \\ [])

get_for_caretaker_q(id, caretaker, opts \\ [])

get_id!(slug)

get_id(slug)

is_stereotype?(acl)

list(opts \\ [])

Lists ACLs we are permitted to see.

list_built_ins()

list_my(user, opts \\ [])

Lists the ACLs we are the registered caretakers of that we are permitted to see. If any are created without permitting the user to see them, they will not be shown.

list_my_q(user, opts \\ [])

query for list_my

list_my_with_counts(user, opts \\ [])

list_q(opts \\ [])

maybe_search(query, text)

opts_for_dropdown()

opts_for_list()

preset_acl_ids()

soft_delete(acl, opts)

Soft-delete the ACL, meaning it will not be displayed anymore, but permissions/grants and controlled information will be preserved. This will not affect objects previously shared with this ACL.

stereotype_ids()

user_default_acl(name)

user_default_acls()