Bonfire.Boundaries.Circles (bonfire_umbrella v0.4.0-classic-beta.28)

Functions

add_to_circles(subject, circles)

changeset(circle \\ %Circle{}, attrs)

changeset(atom, circle, params)

circle_ids(subjects)

circles()

create(user, attrs)

Create a circle for the provided user (and with the user in the circle?)

delete(circle, opts)

Fully delete the circle, including membership and boundary information. This will affect all objects previously shared with members of this circle.

edit(circle, user, params)

get!(slug)

get(slug)

get_for_caretaker(id, caretaker, opts \\ [])

get_for_caretaker_q(id, caretaker, opts \\ [])

query for get

get_id!(slug)

get_id(slug)

get_stereotype_circles(subject, stereotypes)

get_tuple(slug)

is_encircled_by?(subject, circle)

Lists the circles that we are permitted to see.

is_stereotype?(acl)

list_built_ins()

list_by_ids(ids)

list_my(user, opts \\ [])

Lists the circles we are the registered caretakers of that we are permitted to see. If any circles are created without permitting the user to see them, they will not be shown.

list_my_defaults(user \\ nil)

list_my_q(user, opts \\ [])

query for list_my

list_my_with_counts(user, opts \\ [])

list_my_with_global(user, opts \\ [])

list_q(user, opts \\ [])

query for list_visible

list_visible(user, opts \\ [])

Lists the circles that we are permitted to see.

list_visible_q(user, opts \\ [])

query for list_visible

maybe_search(query, text)

remove_from_circles(subject, circles)

stereotype_ids()

to_circle_ids(subjects)