Bonfire.Common.Enums (bonfire_umbrella v0.4.0-classic-beta.68)

Missing functions from Enum

Functions

attr_get_id(attrs, field_name)

count_where(collection, function)

deep_merge(left, right, opts \\ [])

recursively merge structs, maps or lists (into a struct or map)

deep_merge_reduce(list_or_map, opts \\ [])

elem_or(verb, index, fallback)

enum_get(map, key, fallback)

Attempt geting a value out of a map by atom key, or try with string key, or return a fallback

enum_maybe_apply(map, fun, args)

filter_empty(val, fallback)

flatter(list)

Flattens a list by recursively flattening the head and tail of the list

group(list, fun)

Like group_by, except children are required to be unique (will throw otherwise!) and the resulting map does not wrap each item in a list

group_map(list, fun)

has_duplicates?(list, fun \\ nil)

id(id)

input_to_atoms(data, discard_unknown_keys \\ true, including_values \\ false)

map_filter_empty(data)

map_key_replace(map, key, new_key, new_value \\ nil)

Rename a key in a map

map_key_replace_existing(map, key, new_key, new_value \\ nil)

map_new(data, fallback_key \\ :data)

maybe_append(list, value)

@spec maybe_append([any()], any()) :: [any()]

Append an item to a list if it is not nil

maybe_flatten(list)

maybe_from_struct(obj)

maybe_get(map, key, fallback \\ nil)

maybe_list(val, change_fn)

maybe_merge_to_struct(first, precedence)

maybe_put(map, key, value)

conditionally update a map

maybe_to_keyword_list(obj, recursive \\ false)

Converts an enumerable to a list recursively Note: make sure that all keys are atoms, i.e. using input_to_atoms first

maybe_to_map(obj, recursive \\ false)

maybe_to_struct(obj, type \\ nil)

maybe_to_structs(v)

merge_as_map(left, right, opts \\ [])

merge maps or lists (into a map)

merge_changesets(cs1, cs2)

merge_keeping_only_first_keys(map_1, map_2)

merge_structs_as_map(target, merge)

nested_structs_to_maps(struct)

put_new_in(map, list, val)

stringify_keys(map, recursive \\ false)

Convert map atom keys to strings

struct_from_map(a_map, list)

struct_to_map(struct)

uniq_by_id(list)