Bonfire.Common.Errors (bonfire_umbrella v0.9.0-classic-beta.15)

Functions

debug_banner_with_trace(kind, exception, stacktrace, opts \\ [])

debug_exception(msg, exception \\ nil, stacktrace \\ nil, kind \\ :error, opts \\ [])

debug_log(msg, exception \\ nil, stacktrace \\ nil, kind \\ :error)

error_msg(errors)

Turns various kind of errors into an error message string. Used to format errors in a way that can be easily read by the user.

format_banner(kind, exception, stacktrace \\ [], opts \\ [])

Normalizes and formats any throw/error/exit. The message is formatted and displayed in the same format as used by Elixir's CLI. The third argument is the stacktrace which is used to enrich a normalized error with more information. It is only used when the kind is an error.

format_location(opts)

format_mfa(module, fun, arity)

Receives a module, fun and arity and formats it as shown in stacktraces. The arity may also be a list of arguments.

Examples

iex> Exception.format_mfa(Foo, :bar, 1)
{Foo, :bar, "Foo.bar/1"}
iex> Exception.format_mfa(Foo, :bar, [])
{Foo, :bar, "Foo.bar()"}
iex> Exception.format_mfa(nil, :bar, [])
{nil, :bar, "nil.bar()"}

Anonymous functions are reported as -func/arity-anonfn-count-, where func is the name of the enclosing function. Convert to "anonymous fn in func/arity"

format_stacktrace(trace \\ nil, opts)

Formats the stacktrace. A stacktrace must be given as an argument. If not, the stacktrace is retrieved from Process.info/2.

format_stacktrace_entry(entry, opts \\ [])

Receives a stacktrace entry and formats it into a string.

last_stacktrace()

map_error(other, fun)

maybe_ok_error(other, change_fn)

@spec maybe_ok_error(any(), any()) :: any()

Applies change_fn if the first parameter is an {:ok, val} tuple, else returns the value

mf_maybe_link_to_code(text \\ nil, mod, fun, opts)

module_maybe_link_to_code(text \\ nil, mod, opts)

replace_error(other, value)