Bonfire.Common.Extend (bonfire_umbrella v0.9.8-cooperation-beta.26)

Functions

application_loaded?(extension)

disabled_value?(value)

extend_module(module)
(macro)

Extend a module (i.e. define defdelegate and defoverridable for all functions from the source module in the current module. Usage: import Module.Extend extend_module Common.Text

extension_enabled?(module_or_otp_app, opts \\ [])

Whether an Elixir module or extension / OTP app is present AND not part of a disabled Bonfire extension (by having in config something like config :bonfire_common, modularity: :disabled)

extension_loaded?(module_or_otp_app)

Whether an Elixir module or extension / OTP app is present

function_ast(module, fun)

function_code(module, fun)

function_line_number(module, fun)

Return the number of the first line where a function is defined in a module

function_line_numbers(module, fun)

Return the numbers (as a tuple) of the first and last lines of a function's definition in a module

generate_reverse_router!()

has_extension_config?(module_or_otp_app)

Whether an Elixir module or extension / OTP app has configuration keys set up

import_if_enabled(module, fallback_module \\ nil)
(macro)

inject_function(module, fun, target_module \\ nil)

Copy the code defining a function from its original module to one that extends it (or a manually specified module). Usage: Module.Extend.inject_function(Common.TextExtended, :blank?)

macro_inspect(fun)

maybe_extension_loaded(module_or_otp_app)

maybe_extension_loaded!(module_or_otp_app)

maybe_module(module, opts \\ [])

Given an Elixir module, this returns an Elixir module, as long as the module and extension / OTP app is present AND not disabled (eg. by having in config something like config Bonfire.Common.Text, modularity: :disabled)

Important note: you should make sure to use the returned module, rather than the one provided as argument, as it can be different, this allows for swapping out modules in config or user settings (eg. by having in config something like config Bonfire.Common.Text, modularity: MyCustomExtension.Text)

maybe_module!(module, opts \\ [])

maybe_module_loaded(module)

maybe_schema_or_pointer(schema_module)

module_code(module)

module_enabled?(module, opts \\ [])

Whether an Elixir module or extension / OTP app is present AND not disabled (eg. by having in config something like config :bonfire_common, modularity: :disabled)

module_exists?(module)

module_exists_or(module, fallback)

module_file(module)

module_file_code(module)

quoted_import_if_enabled(module, fallback_module \\ nil, caller \\ nil)

quoted_require_if_enabled(module, fallback_module \\ nil, caller \\ nil)

quoted_use_if_enabled(module, fallback_module \\ nil, caller \\ nil)

require_if_enabled(module, fallback_module \\ nil)
(macro)

split_lines(string)

use_if_enabled(module, fallback_module \\ nil)
(macro)