Bonfire.Common.ExtensionBehaviour behaviour (bonfire_umbrella v0.4.0-classic-beta.29)

A Global cache of known Behaviours in Bonfire

Use of the ExtensionBehaviour Service requires ddding @behaviour Bonfire.Common.ExtensionBehaviour in your behaviour modules. This modules when then index those behaviours and all the modules that implement those behaviours at startup.

While this module is a GenServer, it is only responsible for setup of the cache and then exits with :ignore having done so. It is not recommended to restart the service as this will lead to a stop the world garbage collection of all processes and the copying of the entire cache to each process that has queried it since its last local garbage collection.

Callbacks

modules()

@callback modules() :: any()

List modules that implement a behaviour

Functions

apps_to_scan()

behaviour_app_modules(behaviour, behaviours \\ nil)

behaviour_modules(behaviour, behaviours \\ nil)

behaviours()

cached_behaviours()

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

find_adopters_of_behaviours(behaviours \\ find_behaviours())

linked_modules(modules, fun)

module_behaviours(module)

populate()

start_link(_)

@spec start_link(ignored :: term()) :: GenServer.on_start()

Populates the global cache with config_module data via introspection.