Bonfire.Common.Pointers.Tables (bonfire_umbrella v0.9.8-cooperation-beta.5)

Functions

list_ids()

list_schemas()

list_tables(source \\ :code)

Lists all Pointable Tables

list_tables_debug()

many(filters \\ [])

one(id)

one!(filters)

schema_or_table!(schema_or_tablename_or_id)

table!(schema_or_tablename_or_id)

@spec table!(Pointers.Pointer.t()) :: Pointers.Table.t()

Retrieves the Table that a pointer points to Note: Throws an error if the table cannot be found

table_fields(table)