Bonfire.Common.PubSub (bonfire_umbrella v0.9.8-cooperation-beta.5)

Functions

broadcast(topics, payload)

Broadcast some data for realtime updates, for example to a feed or thread

broadcast_with_telemetry(topic, message)
(macro)

Broadcast while attaching telemetry info. The receiving module must use Bonfire.Common.PubSub to correctly unwrap the Event

broadcast_with_telemetry(topic, message, source)

current_function(env)

subscribe(topics, socket)

Subscribe to something for realtime updates, like a feed or thread