Bonfire.Common.SchemaModule behaviour (bonfire_umbrella v0.3.6-classic-beta.7)

Find a context or query module via its schema, backed by a global cache of known schema modules to be queried by their schema, or vice versa (eg. via ContextModule).

Callbacks

context_module()

@callback context_module() :: atom()

Points to the related context module

query_module()

@callback query_module() :: atom()

Points to the related query module

schema_module()

@callback schema_module() :: any()

Declares a schema module

Functions

app_modules()

linked_context_modules()

linked_query_modules()

modules()

@spec modules() :: [atom()]

Callback implementation for Bonfire.Common.ExtensionBehaviour.modules/0.