Bonfire.Common.Settings (bonfire_umbrella v0.9.8-cooperation-beta.26)

Functions

get(key, default \\ nil, opts \\ [])

Get config value for a config key (optionally from a specific OTP app or Bonfire extension)

These two calls have the same result (i.e. specifying a module as the first key will add the OTP app of that module as the first key): get([:bonfire_me, Bonfire.Me.Users]) get(Bonfire.Me.Users)

Same with these two (i.e. not specifying a module or app as the first key will default to the main OTP app): get([:random_atom, :sub_key]) get([:bonfire, :random_atom, :sub_key])

get!(key, opts)

load_instance_settings()

map_put_in(root \\ %{}, keys, value)

put(keys, value, opts)

Put a setting using a key like :key or list of nested keys like [:top_key, :sub_key]

reset_all()

reset_instance()

set(attrs, opts)

Set several settings at once. Accepts nested attributes as map with string keys (which are transformed into keyword list), or a keyword list. Determines what scope to use & sets/updates settings for it.