Bonfire.Common.Types (bonfire_umbrella v0.4.0-classic-beta.77)

Functions

all_object_type_names()

Outputs the names all object types, for the purpose of adding to the localisation strings, as long as the output is piped through to localise_strings/1 at compile time.

defines_struct?(module)

is_numeric(str)

is_ulid?(str)

maybe_convert_ulids(list)

maybe_to_atom!(str)

maybe_to_atom(str)

maybe_to_module(str, force \\ true)

maybe_to_snake(string)

maybe_to_snake_atom(string)

maybe_to_string(atom)

module_to_human_readable(module)

module_to_str(str)

object_type(object)

object_type_display(object_type)

table_id(schema)

table_type(type)

table_types(types)

to_number(str)

typeof(v)

ulid!(object)

ulid(input)

ulids(objects)