Bonfire.Common.Utils (bonfire_umbrella v0.4.0-classic-beta.80)

Functions

apply_error(error, args)

current_account(list)

current_account_and_or_user_ids(arg1)

current_user(current_user_or_socket_or_opts, recursing \\ false)

current_user_id(current_user_or_socket_or_opts, recursing \\ false)

current_user_required!(context)

e(val, fallback)

Returns a value, or a fallback if nil/false

e(object, key, fallback)

Returns a value from a map, or a fallback if not present

e(object, key1, key2, fallback)

Returns a value from a nested map, or a fallback if not present

e(object, key1, key2, key3, fallback)

e(object, key1, key2, key3, key4, fallback)

e(object, key1, key2, key3, key4, key5, fallback)

e(object, key1, key2, key3, key4, key5, key6, fallback)

empty?(v)

maybe(val, change_fn)

Applies change_fn if the first parameter is not nil.

maybe_apply(module, fun, args \\ [], fallback_fun \\ &apply_error/2)

Helpers for calling hypothetical functions in other modules

maybe_from_opts(opts, key, fallback \\ nil)

ok_unwrap(val, fallback \\ nil)

replace_nil(other, value)

socket_connected?(socket)

to_options(user_or_socket_or_opts)