Bonfire.Encrypt.Secret (bonfire_umbrella v0.4.0-classic-beta.71)

A mixin that stores [metadata about] encrypted secrets

Functions

burn!(secret, event_extra \\ [])

count_secrets()

do_burn!(secret)

Burns a secret or throws Burned secrets have no iv and no ciphertext

get_secret!(id)

Reads secret with id or throws

get_secret(id)

Reads secret with id or returns error

insert!(presecret_attrs)

Inserts secret or throws presecret_attrs is a map of attrs from the Presecret struct. We transform this into fields on the Secret. Easier to send base64 to to the browser with Presecret and store raw binary in the Secret.

new()

topic(id)

validate_content_size(changeset)

validate_iv_size(changeset)