Bonfire.Social.Activities (bonfire_umbrella v0.4.0-classic-beta.6)

Functions

activity_preloads(query, opts)

activity_preloads(query, preloads, opts)

activity_under_object(top_object)

activity_under_object(activity, object)

all_verb_names()

Outputs the names all object verbs, for the purpose of adding to the localisation strings, as long as the output is piped through to localise_strings/1 at compile time.

all_verb_names_extra()

as_permitted_for(q, opts \\ [], verbs \\ [:see, :read])

assigns_with_object_under_activity(assigns)

build_assoc(thing, verb, subject)

build_assoc(thing, verb, subject, object)

by_subject_verb_object_q(subject, verb, object)

cast(changeset, verb, creator, opts)

create(subject, verb, object, activity_id \\ nil)

Create an Activity NOTE: you will usually want to use cast/3 instead

delete(filters)

Delete activities, using specific filters

delete_by_subject_verb_object(subject, verb, object)

Delete an activity (usage by things like unlike)

get(id, opts)

Get an activity by its ID

load_object(id_or_pointer, opts \\ [])

object_from_activity(activity)

object_preload_activity(object, verb \\ :create)

object_preload_create_activity(object)

object_under_activity(activity, object)

put_assoc(changeset, verb, subject)

put_assoc(changeset, verb, subject, object)

query(filters \\ [], opts_or_current_user \\ [])

query_object_preload_activity(q, verb \\ :create, object_id_field \\ :id, opts \\ [])

query_object_preload_create_activity(q, opts \\ [])

read(query, opts \\ [])

Get an activity by its object ID (usually a create activity)

reply_to_as_permitted_for(q, opts \\ [], verbs \\ [:see, :read])

schema_module()

Callback implementation for Bonfire.Common.ContextModule.schema_module/0.

verb_congugate(verb)

verb_display(verb)

verb_id(verb)

verb_maybe_modify(verb, activity)

verb_name(slug)