Bonfire.Social.Edges (bonfire_umbrella v0.9.8-cooperation-beta.26)

Functions

changeset(schema, subject, verb, object, options)

changeset_base(schema, subject, object, options)

changeset_extra(schema, subject, verb, object, options)

changeset_without_caretaker(schema, subject, verb, object, options)

count(type, filters_or_object, opts \\ [])

count_for_subject(type, subject, object, opts)

delete_by_both(me, schema, object)

delete_by_object(user)

delete_by_subject(user)

edge_module_query(schema_or_context, args)

exists?(schema_or_context, subject, object, opts)

filters_from_opts(assigns)

get(schema_or_context, subject, object, opts \\ [])

get!(schema_or_context, subject, objects, opts \\ [])

insert(changeset, subject \\ nil, object \\ nil)

insert(schema, subject, verb, object, options)

last(schema_or_context, subject, object, opts)

retrieves the last edge of a given type, subject, and object from the database

last_date(type, subject, object, opts)

retrieves the date of the last edge of a given type, subject, and object from the database

put_edge_assoc(changeset, subject, object)

put_edge_assoc(changeset, schema, subject, object)

query(filters, opts)

query_parent(query_schema, filters, opts)