Bonfire.Social.Flags (bonfire_umbrella v0.9.8-cooperation-beta.30)

Functions

ap_publish_activity(subject, verb, flag)

ap_receive_activity(creator, activity, objects)

by_flagged(object)

by_flagger(subject)

federation_module()

Callback implementation for Bonfire.Federate.ActivityPub.FederationModules.federation_module/0.

flag(flagger, flagged, opts \\ [])

flagged?(user, object)

get(subject, object, opts \\ [])

get!(subject, object, opts \\ [])

instance_moderators()

list(opts)

list_by(by_user, opts \\ [])

List flags by the user and which are in their outbox

list_my(opts)

List current user's flags, which are in their outbox

list_of(object, opts \\ [])

List flag of an object and which are in a feed

list_paginated(filters, opts)

list_preloaded(opts)

moderators(object)

query(filters, opts)

query_module()

Callback implementation for Bonfire.Common.ContextModule.query_module/0.

schema_module()

Callback implementation for Bonfire.Common.ContextModule.schema_module/0.

unflag(flagger, flagged)