Bonfire.Social.Follows (bonfire_umbrella v0.4.0-classic-beta.16)

Functions

accept(request, opts)

Accepts a follow request, publishes to feeds and federates. Parameter are a Request (or its ID) plus the subject as current_user

accept_from(subject, opts)

Accepts a follow request, publishes to feeds and federates. Parameters are the requester plus the subject as current_user

all_by_object(user, opts \\ [])

all_by_subject(user, opts \\ [])

all_followed_outboxes(user, opts \\ [])

all_objects_by_subject(user, opts \\ [])

all_subjects_by_object(user, opts \\ [])

ap_publish_activity(subject, verb, object)

ap_receive_activity(follower, activity, object)

federation_module()

follow(user, object, opts \\ [])

Follow someone/something. In case of success, publishes to feeds and federates.

If the user is not permitted to follow the object, or the object is a remote actor, will instead request to follow.

following?(subject, object)

get!(subject, object, opts \\ [])

get(subject, object, opts \\ [])

ignore(request, opts)

list_followed(user, opts \\ [])

list_followers(user, opts \\ [])

list_my_followed(current_user, opts \\ [])

list_my_followers(current_user, opts \\ [])

query(filters, opts)

requested?(subject, object)

schema_module()

Callback implementation for Bonfire.Common.ContextModule.schema_module/0.

unfollow(user, object)