Bonfire.Social.Objects (bonfire_umbrella v0.9.8-cooperation-beta.11)

Functions

ap_maybe_delete(creator, object)

ap_receive_activity(creator, activity, object)

as_permitted_for(q, opts \\ [], verbs \\ [:see, :read])

cast(changeset, attrs, creator, opts)

Handles casting:

Creator

Caretaker

Threaded replies (when present)

Tags/Mentions (when present)

Acls

Activity

cast_acl(changeset, creator, opts)

cast_basic(changeset, attrs, creator, opts)

Handles casting:

Creator

Caretaker

Acls

cast_caretaker(changeset, caretaker)

cast_creator(changeset, creator)

cast_creator_caretaker(changeset, user)

cast_mini(changeset, attrs, creator, opts)

Handles casting:

Acls

cast_publish(changeset, attrs, creator, opts)

Handles casting:

Acls

Activity

Feed Publishes

cast_sensitivity(changeset, sensitive?)

changeset_named(object \\ %{}, attrs)

delete(object, opts)

delete_apply_error(error, args)

do_delete(objects, opts)

federation_module()

list_query(type_or_query \\ nil, opts)

maybe_generic_delete(type, object, options \\ [])

maybe_preload_activity_object(pointer, current_user)

object_creator(object)

preload_creator(object)

preload_reply_creator(object)

publish(creator, verb, thing, opts_or_attrs \\ nil, for_module \\ __MODULE__)

query_base(type \\ nil)

Returns a basic query over undeleted pointable objects in the system, optionally limited to one or more types.

read(object_id, socket_or_current_user)

reset_preset_boundary(creator, thing, previous_preset, opts_or_attrs \\ [], for_module \\ nil)

run_epic(type, options \\ [], on \\ :object)

set_boundaries(creator, thing, opts_or_attrs \\ [], for_module \\ __MODULE__)

set_name(id, name, opts)

set_sensitivity(pointer, arg2)