Bonfire.Social.PostContents (bonfire_umbrella v0.9.8-cooperation-beta.5)

Query and manipulate post contents

Functions

all_text_content(attrs)

ap_prepare_attrs(creator, activity_data, post_data, direct_recipients \\ [])

ap_receive_activity(creator, activity_data, object)

cast(changeset, attrs, creator, boundary, opts)

Given a changeset, post content attributes, creator, boundary and options, returns a changeset prepared with relevant attributes and associations

changeset(cs \\ %PostContent{}, attrs)

edit(current_user, id, attrs)

editor_output_content_type(user)

federation_module()

get(id, opts \\ [])

Given a post content ID, returns the corresponding post content

get_versions(post_content)

indexing_object_format(obj)

maybe_detect_languages(attrs, fields \\ [:name, :summary, :html_body])

maybe_prepare_contents(attrs, creator, boundary, opts)

Given post content attributes, creator, boundary, and options, prepares the post contents for processing by detecting languages, mentions, hashtags, and urls.

merge_with_body_or_nil(attrs, map)

no_known_output(error, args)

one(filters, opts \\ [])

Given a set of filters, returns a single post content matching those filters

prepare_text(text, arg2, opts)

query(filters, opts \\ [])

Given a set of filters, returns an Ecto.Query for matching post contents.

query_base()