Bonfire.Social.Seen (bonfire_umbrella v0.9.8-cooperation-beta.5)

Track seen/unseen (similar to read/unread, but only indicates that it was displayed in a feed or other listing for the user, not that they actually read it) status of things (usually Activities)

Functions

get(subject, object, opts \\ [])

get!(subject, object, opts \\ [])

mark_seen(subject, object)

mark_unseen(subject, object)

query(filters, opts)

schema_module()

Callback implementation for Bonfire.Common.ContextModule.schema_module/0.

seen?(user, object)