Bonfire.Social.Threads (bonfire_umbrella v0.9.8-cooperation-beta.21)

Functions

ap_prepare(object_or_thread_or_reply_to_id, key \\ :thread_id)

Key can be for eg. :thread_id or :reply_to_id

arrange_replies(replies, opts \\ [])

arrange_replies_tree(replies, opts \\ [])

base_query()

cast(changeset, attrs, user, preset_or_custom_boundary)

Handles casting related to the reply and threading. If it's not a reply or the user is not permitted to reply to the thing, a new thread will be created.

changeset(replied \\ %Replied{}, attrs)

count_participants(thread_id, opts \\ [])

Count participants in a thread (depending on user's boundaries)

fetch_participants(thread_id, opts \\ [])

List participants in a thread (depending on user's boundaries)

filter(arg1, thread_id, query)

Group per-thread

find_reply_to(attrs, user)

Returns {:ok, reply} or {:error, reason}, where reason may be:

:not_found - we could not find the object you are replying to.

:not_permitted - you do not have permission to reply to this.

find_thread(attrs, user)

list_participants(activity_or_object, thread_or_object_id \\ nil, opts \\ [])

List participants of an activity or thread (depending on user's boundaries)

list_replies(thread, opts \\ [])

mark_all_seen(filters, opts)

maybe_re_order_result(result, opts)

re-order distinct threads after DISTINCT ON ordered them by thread_id - Note: does not support pagination

query(filter, opts)

re_order_using_subquery(query, opts)

re-order distinct threads after DISTINCT ON ordered them by thread_id - Note: this results in (Ecto.QueryError) cannot preload associations in subquery in query

read(object_id, socket_or_current_user)

schema_module()

Callback implementation for Bonfire.Common.QueryModule.schema_module/0.

unseen_count(filters, opts)

unseen_query(filters, opts)