Bonfire.Tag.Tags (bonfire_umbrella v0.9.8-cooperation-beta.18)

Functions

find(id, types \\ nil)

get(id, opts \\ [])

indexing_object_format(object)

indexing_object_format_name(object)

list_trending(in_last_x_days \\ 30, limit \\ 10)

list_trending_reset(in_last_x_days \\ 30, limit \\ 10)

many(filters \\ [], opts \\ [])

Retrieves a list of tags by arbitrary filters. Used by:

Various parts of the codebase that need to query for tags (inc. tests)

maybe_find_tag(current_user, id_or_username_or_url, types \\ nil)

Try to find one (best-match) tag

maybe_find_tags(current_user, id_or_username_or_url, types \\ nil)

Search / autocomplete for tags by name

maybe_lookup_tag(id_or_username_or_url, prefix \\ "@")

Lookup a single for a tag by its name/username

maybe_tag(user, thing, tags \\ nil, boost_category_mentions? \\ true)

Maybe tag something

maybe_taxonomy_tag(user, id)

one(filters, opts \\ [])

Retrieves a single tag by arbitrary filters. Used by:

GraphQL Item queries

ActivityPub integration

Various parts of the codebase that need to query for tags (inc. tests)

tag_ids(tags)

tag_something(user, thing, tags, boost_category_mentions? \\ true)

tag existing thing with one or multiple Tags, Pointers, or anything that can be made into a tag