Bonfire.UI.Common.PreviewContentLive (bonfire_umbrella v0.9.8-cooperation-beta.2)

A special 'modal' for previewing objects from a feed without redirecting to a new view (so we can go back to where we were in the feed)

Properties

id :string, required: true - The id of the live component (required by LiveView for stateful components).

source_inspector_attrs :map, default: %{}

title_text :string, default: nil - The title of the modal. Only used if no title slot is passed.

title_class :css_class, default: nil - The classes of the title of the modal

show :boolean, default: false - Force modal to be open

no_header :boolean, default: false - Optional prop to hide the header at the top of the modal

modal_class :string, default: nil - The classes of the modal.

modal_assigns :any, default: [] - Additional assigns for the modal

opts :keyword, default: [] - Additional attributes to add onto the modal wrapper

Slots

default, arg: [:autocomplete] - Slots for the contents of the modal, title, buttons...

open_btn

action_btns

cancel_btn

title

Functions

(arg, modifiers)
(macro)

alert(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.alert/1.

alert(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.alert/2.

alert(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.alert/3.

allows_hit_testing(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.allows_hit_testing/1.

allows_hit_testing(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.allows_hit_testing/2.

animation(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.animation/1.

animation(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.animation/2.

animation(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.animation/3.

antialiased(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.antialiased/1.

antialiased(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.antialiased/2.

aspect_ratio(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.aspect_ratio/1.

aspect_ratio(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.aspect_ratio/2.

aspect_ratio(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.aspect_ratio/3.

autocorrection_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.autocorrection_disabled/1.

autocorrection_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.autocorrection_disabled/2.

background(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.background/1.

background(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.background/2.

background(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.background/3.

background_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.background_style/1.

background_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.background_style/2.

badge(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.badge/1.

badge(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.badge/2.

baseline_offset(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.baseline_offset/1.

baseline_offset(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.baseline_offset/2.

blend_mode(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.blend_mode/1.

blend_mode(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.blend_mode/2.

blur()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.blur/0.

blur(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.blur/1.

blur(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.blur/2.

bold(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.bold/1.

bold(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.bold/2.

border(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.border/1.

border(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.border/2.

border(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.border/3.

brightness(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.brightness/1.

brightness(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.brightness/2.

button_border_shape(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.button_border_shape/1.

button_border_shape(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.button_border_shape/2.

button_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.button_style/1.

button_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.button_style/2.

clip_shape(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.clip_shape/1.

clip_shape(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.clip_shape/2.

clip_shape(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.clip_shape/3.

clipped()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.clipped/0.

clipped(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.clipped/1.

clipped(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.clipped/2.

color_invert()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.color_invert/0.

color_invert(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.color_invert/1.

color_invert(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.color_invert/2.

color_multiply(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.color_multiply/1.

color_multiply(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.color_multiply/2.

compositing_group()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.compositing_group/0.

compositing_group(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.compositing_group/1.

compositing_group(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.compositing_group/2.

confirmation_dialog(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.confirmation_dialog/1.

confirmation_dialog(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.confirmation_dialog/2.

confirmation_dialog(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.confirmation_dialog/3.

container_shape(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.container_shape/1.

container_shape(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.container_shape/2.

content_shape(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.content_shape/1.

content_shape(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.content_shape/2.

content_shape(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.content_shape/3.

content_shape(arg1, arg2, arg3, arg4)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.content_shape/4.

content_transition(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.content_transition/1.

content_transition(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.content_transition/2.

context_menu()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.context_menu/0.

context_menu(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.context_menu/1.

context_menu(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.context_menu/2.

contrast(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.contrast/1.

contrast(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.contrast/2.

control_group_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.control_group_style/1.

control_group_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.control_group_style/2.

control_size(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.control_size/1.

control_size(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.control_size/2.

coordinate_space(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.coordinate_space/1.

coordinate_space(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.coordinate_space/2.

corner_radius(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.corner_radius/1.

corner_radius(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.corner_radius/2.

corner_radius(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.corner_radius/3.

date_picker_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.date_picker_style/1.

date_picker_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.date_picker_style/2.

default_wheel_picker_item_height(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.default_wheel_picker_item_height/1.

default_wheel_picker_item_height(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.default_wheel_picker_item_height/2.

defers_system_gestures(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.defers_system_gestures/1.

defers_system_gestures(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.defers_system_gestures/2.

delete_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.delete_disabled/1.

delete_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.delete_disabled/2.

digital_crown_accessory()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.digital_crown_accessory/0.

digital_crown_accessory(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.digital_crown_accessory/1.

digital_crown_accessory(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.digital_crown_accessory/2.

digital_crown_accessory_visibility(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.digital_crown_accessory_visibility/1.

digital_crown_accessory_visibility(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.digital_crown_accessory_visibility/2.

disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.disabled/1.

disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.disabled/2.

disclosure_group_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.disclosure_group_style/1.

disclosure_group_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.disclosure_group_style/2.

do_handle_event(binary, _, socket)

draggable(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.draggable/1.

draggable(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.draggable/2.

draggable(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.draggable/3.

drawing_group()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.drawing_group/0.

drawing_group(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.drawing_group/1.

drawing_group(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.drawing_group/2.

drop_destination()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.drop_destination/0.

drop_destination(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.drop_destination/1.

drop_destination(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.drop_destination/2.

dynamic_type_size(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.dynamic_type_size/1.

dynamic_type_size(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.dynamic_type_size/2.

fill(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.fill/1.

fill(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.fill/2.

fill(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.fill/3.

find_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.find_disabled/1.

find_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.find_disabled/2.

find_navigator()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.find_navigator/0.

find_navigator(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.find_navigator/1.

find_navigator(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.find_navigator/2.

fixed_size()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.fixed_size/0.

fixed_size(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.fixed_size/1.

fixed_size(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.fixed_size/2.

flips_for_right_to_left_layout_direction(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.flips_for_right_to_left_layout_direction/1.

flips_for_right_to_left_layout_direction(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.flips_for_right_to_left_layout_direction/2.

focus_scope(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focus_scope/1.

focus_scope(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focus_scope/2.

focus_section()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focus_section/0.

focus_section(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focus_section/1.

focus_section(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focus_section/2.

focusable(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focusable/1.

focusable(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.focusable/2.

font(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font/1.

font(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font/2.

font_design(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font_design/1.

font_design(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font_design/2.

font_weight(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font_weight/1.

font_weight(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font_weight/2.

font_width(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font_width/1.

font_width(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.font_width/2.

foreground_color(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.foreground_color/1.

foreground_color(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.foreground_color/2.

foreground_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.foreground_style/1.

foreground_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.foreground_style/2.

foreground_style(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.foreground_style/3.

foreground_style(arg1, arg2, arg3, arg4)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.foreground_style/4.

frame()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.frame/0.

frame(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.frame/1.

frame(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.frame/2.

full_screen_cover()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.full_screen_cover/0.

full_screen_cover(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.full_screen_cover/1.

full_screen_cover(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.full_screen_cover/2.

gauge_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.gauge_style/1.

gauge_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.gauge_style/2.

gesture(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.gesture/1.

gesture(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.gesture/2.

gesture(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.gesture/3.

grayscale(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grayscale/1.

grayscale(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grayscale/2.

grid_cell_anchor(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_cell_anchor/1.

grid_cell_anchor(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_cell_anchor/2.

grid_cell_columns(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_cell_columns/1.

grid_cell_columns(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_cell_columns/2.

grid_cell_unsized_axes(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_cell_unsized_axes/1.

grid_cell_unsized_axes(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_cell_unsized_axes/2.

grid_column_alignment(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_column_alignment/1.

grid_column_alignment(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.grid_column_alignment/2.

group_box_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.group_box_style/1.

group_box_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.group_box_style/2.

handle_event(action, attrs, socket)

Callback implementation for Phoenix.LiveComponent.handle_event/3.

handle_info(info, socket)

header_prominence(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.header_prominence/1.

header_prominence(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.header_prominence/2.

help(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.help/1.

help(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.help/2.

hidden()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hidden/0.

hidden(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hidden/1.

hidden(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hidden/2.

horizontal_radio_group_layout()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.horizontal_radio_group_layout/0.

horizontal_radio_group_layout(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.horizontal_radio_group_layout/1.

horizontal_radio_group_layout(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.horizontal_radio_group_layout/2.

hover_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hover_effect/1.

hover_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hover_effect/2.

hue_rotation(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hue_rotation/1.

hue_rotation(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.hue_rotation/2.

ignores_safe_area(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.ignores_safe_area/1.

ignores_safe_area(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.ignores_safe_area/2.

ignores_safe_area(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.ignores_safe_area/3.

image_scale(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.image_scale/1.

image_scale(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.image_scale/2.

index_view_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.index_view_style/1.

index_view_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.index_view_style/2.

inset(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.inset/1.

inset(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.inset/2.

interactive_dismiss_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.interactive_dismiss_disabled/1.

interactive_dismiss_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.interactive_dismiss_disabled/2.

interpolation(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.interpolation/1.

interpolation(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.interpolation/2.

italic(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.italic/1.

italic(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.italic/2.

kerning(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.kerning/1.

kerning(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.kerning/2.

keyboard_shortcut(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyboard_shortcut/1.

keyboard_shortcut(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyboard_shortcut/2.

keyboard_shortcut(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyboard_shortcut/3.

keyboard_type(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyboard_type/1.

keyboard_type(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyboard_type/2.

keyframe_animator()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyframe_animator/0.

keyframe_animator(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyframe_animator/1.

keyframe_animator(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.keyframe_animator/2.

label_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.label_style/1.

label_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.label_style/2.

labels_hidden()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.labels_hidden/0.

labels_hidden(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.labels_hidden/1.

labels_hidden(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.labels_hidden/2.

layout_priority(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.layout_priority/1.

layout_priority(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.layout_priority/2.

line_limit(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.line_limit/1.

line_limit(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.line_limit/2.

line_spacing(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.line_spacing/1.

line_spacing(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.line_spacing/2.

list_item_tint(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_item_tint/1.

list_item_tint(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_item_tint/2.

list_row_background(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_background/1.

list_row_background(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_background/2.

list_row_insets(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_insets/1.

list_row_insets(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_insets/2.

list_row_separator(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_separator/1.

list_row_separator(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_separator/2.

list_row_separator(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_separator/3.

list_row_separator_tint(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_separator_tint/1.

list_row_separator_tint(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_separator_tint/2.

list_row_separator_tint(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_row_separator_tint/3.

list_section_separator(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_section_separator/1.

list_section_separator(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_section_separator/2.

list_section_separator(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_section_separator/3.

list_section_separator_tint(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_section_separator_tint/1.

list_section_separator_tint(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_section_separator_tint/2.

list_section_separator_tint(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_section_separator_tint/3.

list_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_style/1.

list_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.list_style/2.

luminance_to_alpha()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.luminance_to_alpha/0.

luminance_to_alpha(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.luminance_to_alpha/1.

luminance_to_alpha(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.luminance_to_alpha/2.

mask()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.mask/0.

mask(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.mask/1.

mask(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.mask/2.

matched_geometry_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.matched_geometry_effect/1.

matched_geometry_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.matched_geometry_effect/2.

menu_action_dismiss_behavior(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_action_dismiss_behavior/1.

menu_action_dismiss_behavior(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_action_dismiss_behavior/2.

menu_indicator(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_indicator/1.

menu_indicator(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_indicator/2.

menu_order(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_order/1.

menu_order(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_order/2.

menu_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_style/1.

menu_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.menu_style/2.

minimum_scale_factor(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.minimum_scale_factor/1.

minimum_scale_factor(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.minimum_scale_factor/2.

monospaced(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.monospaced/1.

monospaced(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.monospaced/2.

monospaced_digit()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.monospaced_digit/0.

monospaced_digit(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.monospaced_digit/1.

mount(socket)

Callback implementation for Phoenix.LiveComponent.mount/1.

move_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.move_disabled/1.

move_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.move_disabled/2.

multiline_text_alignment(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.multiline_text_alignment/1.

multiline_text_alignment(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.multiline_text_alignment/2.

navigation_bar_back_button_hidden(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_bar_back_button_hidden/1.

navigation_bar_back_button_hidden(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_bar_back_button_hidden/2.

navigation_bar_title_display_mode(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_bar_title_display_mode/1.

navigation_bar_title_display_mode(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_bar_title_display_mode/2.

navigation_subtitle(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_subtitle/1.

navigation_subtitle(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_subtitle/2.

navigation_title(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_title/1.

navigation_title(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.navigation_title/2.

offset()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.offset/0.

offset(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.offset/1.

offset(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.offset/2.

offset_shape()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.offset_shape/0.

offset_shape(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.offset_shape/1.

offset_shape(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.offset_shape/2.

on_appear()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_appear/0.

on_appear(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_appear/1.

on_appear(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_appear/2.

on_delete_command()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_delete_command/0.

on_delete_command(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_delete_command/1.

on_delete_command(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_delete_command/2.

on_disappear()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_disappear/0.

on_disappear(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_disappear/1.

on_disappear(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_disappear/2.

on_hover()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_hover/0.

on_hover(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_hover/1.

on_hover(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_hover/2.

on_move_command()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_move_command/0.

on_move_command(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_move_command/1.

on_move_command(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_move_command/2.

on_submit(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_submit/1.

on_submit(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_submit/2.

on_submit(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.on_submit/3.

opacity(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.opacity/1.

opacity(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.opacity/2.

overlay(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.overlay/1.

overlay(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.overlay/2.

overlay(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.overlay/3.

padding()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.padding/0.

padding(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.padding/1.

padding(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.padding/2.

padding(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.padding/3.

persistent_system_overlays(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.persistent_system_overlays/1.

persistent_system_overlays(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.persistent_system_overlays/2.

picker_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.picker_style/1.

picker_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.picker_style/2.

popover()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.popover/0.

popover(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.popover/1.

popover(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.popover/2.

position()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.position/0.

position(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.position/1.

position(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.position/2.

preferred_color_scheme(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.preferred_color_scheme/1.

preferred_color_scheme(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.preferred_color_scheme/2.

prefers_default_focus(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.prefers_default_focus/1.

prefers_default_focus(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.prefers_default_focus/2.

prefers_default_focus(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.prefers_default_focus/3.

presentation_background(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_background/1.

presentation_background(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_background/2.

presentation_background_interaction(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_background_interaction/1.

presentation_background_interaction(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_background_interaction/2.

presentation_compact_adaptation()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_compact_adaptation/0.

presentation_compact_adaptation(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_compact_adaptation/1.

presentation_compact_adaptation(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_compact_adaptation/2.

presentation_content_interaction(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_content_interaction/1.

presentation_content_interaction(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_content_interaction/2.

presentation_corner_radius(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_corner_radius/1.

presentation_corner_radius(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_corner_radius/2.

presentation_detents(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_detents/1.

presentation_detents(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_detents/2.

presentation_detents(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_detents/3.

presentation_drag_indicator(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_drag_indicator/1.

presentation_drag_indicator(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.presentation_drag_indicator/2.

privacy_sensitive(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.privacy_sensitive/1.

privacy_sensitive(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.privacy_sensitive/2.

progress_view_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.progress_view_style/1.

progress_view_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.progress_view_style/2.

projection_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.projection_effect/1.

projection_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.projection_effect/2.

redacted()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.redacted/0.

redacted(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.redacted/1.

redacted(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.redacted/2.

refreshable()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.refreshable/0.

refreshable(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.refreshable/1.

refreshable(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.refreshable/2.

rename_action(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rename_action/1.

rename_action(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rename_action/2.

render(assigns)

Callback implementation for Phoenix.LiveComponent.render/1.

render_blank(assigns)

render_native(assigns)

rendering_mode(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rendering_mode/1.

rendering_mode(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rendering_mode/2.

replace_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.replace_disabled/1.

replace_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.replace_disabled/2.

resizable()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.resizable/0.

resizable(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.resizable/1.

resizable(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.resizable/2.

rotation(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation/1.

rotation(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation/2.

rotation(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation/3.

rotation_3d_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation_3d_effect/1.

rotation_3d_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation_3d_effect/2.

rotation_3d_effect(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation_3d_effect/3.

rotation_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation_effect/1.

rotation_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation_effect/2.

rotation_effect(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.rotation_effect/3.

safe_area_inset()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.safe_area_inset/0.

safe_area_inset(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.safe_area_inset/1.

safe_area_inset(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.safe_area_inset/2.

saturation(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.saturation/1.

saturation(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.saturation/2.

scale_effect()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scale_effect/0.

scale_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scale_effect/1.

scale_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scale_effect/2.

scaled_to_fill()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scaled_to_fill/0.

scaled_to_fill(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scaled_to_fill/1.

scaled_to_fill(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scaled_to_fill/2.

scaled_to_fit()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scaled_to_fit/0.

scaled_to_fit(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scaled_to_fit/1.

scaled_to_fit(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scaled_to_fit/2.

scene_padding(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scene_padding/1.

scene_padding(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scene_padding/2.

scene_padding(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scene_padding/3.

scroll_bounce_behavior(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_bounce_behavior/1.

scroll_bounce_behavior(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_bounce_behavior/2.

scroll_bounce_behavior(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_bounce_behavior/3.

scroll_content_background(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_content_background/1.

scroll_content_background(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_content_background/2.

scroll_disabled(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_disabled/1.

scroll_disabled(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_disabled/2.

scroll_dismisses_keyboard(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_dismisses_keyboard/1.

scroll_dismisses_keyboard(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_dismisses_keyboard/2.

scroll_indicators(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_indicators/1.

scroll_indicators(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_indicators/2.

scroll_indicators(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.scroll_indicators/3.

search_completion(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_completion/1.

search_completion(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_completion/2.

search_scopes(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_scopes/1.

search_scopes(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_scopes/2.

search_scopes(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_scopes/3.

search_suggestions(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_suggestions/1.

search_suggestions(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.search_suggestions/2.

searchable()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.searchable/0.

searchable(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.searchable/1.

searchable(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.searchable/2.

shadow()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.shadow/0.

shadow(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.shadow/1.

shadow(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.shadow/2.

sheet()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.sheet/0.

sheet(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.sheet/1.

sheet(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.sheet/2.

sigil_LVN(arg, modifiers)
(macro)

sigil_p(route, extra)
(macro)

sigil_SWIFTUI(arg, modifiers)
(macro)

size()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.size/0.

size(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.size/1.

size(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.size/2.

status_bar_hidden(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.status_bar_hidden/1.

status_bar_hidden(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.status_bar_hidden/2.

strikethrough(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.strikethrough/1.

strikethrough(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.strikethrough/2.

strikethrough(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.strikethrough/3.

stroke(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.stroke/1.

stroke(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.stroke/2.

stroke(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.stroke/3.

stroke_border(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.stroke_border/1.

stroke_border(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.stroke_border/2.

stroke_border(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.stroke_border/3.

submit_label(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.submit_label/1.

submit_label(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.submit_label/2.

submit_scope(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.submit_scope/1.

submit_scope(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.submit_scope/2.

swipe_actions()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.swipe_actions/0.

swipe_actions(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.swipe_actions/1.

swipe_actions(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.swipe_actions/2.

symbol_rendering_mode(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.symbol_rendering_mode/1.

symbol_rendering_mode(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.symbol_rendering_mode/2.

symbol_variant(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.symbol_variant/1.

symbol_variant(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.symbol_variant/2.

tab_item(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tab_item/1.

tab_item(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tab_item/2.

tab_view_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tab_view_style/1.

tab_view_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tab_view_style/2.

table_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.table_style/1.

table_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.table_style/2.

text_case(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_case/1.

text_case(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_case/2.

text_content_type(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_content_type/1.

text_content_type(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_content_type/2.

text_field_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_field_style/1.

text_field_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_field_style/2.

text_input_autocapitalization(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_input_autocapitalization/1.

text_input_autocapitalization(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_input_autocapitalization/2.

text_selection(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_selection/1.

text_selection(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.text_selection/2.

tint(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tint/1.

tint(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tint/2.

toggle_style(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toggle_style/1.

toggle_style(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toggle_style/2.

toolbar()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar/0.

toolbar(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar/1.

toolbar(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar/2.

toolbar_background(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_background/1.

toolbar_background(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_background/2.

toolbar_background(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_background/3.

toolbar_color_scheme(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_color_scheme/1.

toolbar_color_scheme(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_color_scheme/2.

toolbar_color_scheme(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_color_scheme/3.

toolbar_role(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_role/1.

toolbar_role(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_role/2.

toolbar_title_menu()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_title_menu/0.

toolbar_title_menu(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_title_menu/1.

toolbar_title_menu(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_title_menu/2.

toolbar_visibility(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_visibility/1.

toolbar_visibility(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_visibility/2.

toolbar_visibility(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.toolbar_visibility/3.

tracking(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tracking/1.

tracking(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.tracking/2.

transform_effect(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.transform_effect/1.

transform_effect(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.transform_effect/2.

transition(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.transition/1.

transition(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.transition/2.

trim(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.trim/1.

trim(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.trim/2.

truncation_mode(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.truncation_mode/1.

truncation_mode(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.truncation_mode/2.

underline(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.underline/1.

underline(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.underline/2.

underline(arg1, arg2, arg3)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.underline/3.

unredacted()

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.unredacted/0.

unredacted(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.unredacted/1.

unredacted(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.unredacted/2.

update(assigns, socket)

Callback implementation for Phoenix.LiveComponent.update/2.

z_index(arg1)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.z_index/1.

z_index(arg1, arg2)

See Bonfire.UI.Common.PreviewContentLive.LiveViewNativeSwiftUi.z_index/2.