Bonfire.UI.Common.ReusableModalLive (bonfire_umbrella v0.4.0-classic-beta.7)

The classic modal

Properties

id :string, required: true - The id of the live component (required by LiveView for stateful components).

title_text :string, default: nil - The title of the modal. Only used if no title slot is passed.

image_preview :boolean, default: false - If the modal is a preview of an image, set this to true.

title_class :css_class, default: nil - The classes of the title of the modal

modal_class :css_class, default: @modal_class - The classes of the modal.

wrapper_class :css_class, default: nil - The classes of the modal wrapper.

action_btns_wrapper_class :css_class, default: nil - The classes around the action/submit button(s) on the modal

cancel_btn_class :css_class, default: nil - The classes of the close/cancel button on the modal. Only used if no close_btn slot is passed.

cancel_btn_text :string, default: nil

show :boolean, default: false - Force modal to be open

without_form :boolean, default: false

form_opts :any, default: []

form_id :string, default: @form_id - The ID of the form to submit

no_actions :boolean, default: false - Optional prop to hide the actions at the bottom of the modal

no_header :boolean, default: false - Optional prop to hide the header at the top of the modal

opts :keyword, default: [] - Additional attributes to add onto the modal wrapper

autocomplete :list, default: []

Slots

default, args: [:autocomplete] - Slots for the contents of the modal, title, buttons...

open_btn

action_btns

cancel_btn

title

Functions

default_assigns()

do_handle_event(binary, _, socket)

handle_event(action, attrs, socket)

Callback implementation for Phoenix.LiveComponent.handle_event/3.

handle_info(info, socket)

mount(socket)

Callback implementation for Phoenix.LiveComponent.mount/1.

render(assigns)

Callback implementation for Phoenix.LiveComponent.render/1.

sigil_p(route, extra)
(macro)

update(assigns, socket)

Callback implementation for Phoenix.LiveComponent.update/2.