Bonfire.UI.Common.SmartInput.LiveHandler (bonfire_umbrella v0.9.0-classic-beta.15)

Functions

active_smart_input_component(smart_input_component, create_object_type)

active_smart_input_components(smart_input_component, create_object_type)

activity_type_or_reply(assigns)

all_smart_input_components()

close_smart_input(target, js \\ %JS{})

confirm_close_smart_input(target, reusable_modal_id, js \\ %JS{})

handle_event(action, params, socket)

max_length()

maybe_from_json(json)

maybe_from_json_string(json)

open(context, assigns)

Open the composer and set assigns

open_smart_input_with_text_suggestion(text, set_assigns, socket_or_context)

replace_input_next_time(socket_or_context)

reply_to_param(arg1)

reset_input(socket)

select_smart_input(target, component, create_object_type, opts \\ [], js \\ %JS{})

set(context, assigns)

Set assigns in the smart input from anywhere in the app (whether using a live component or sticky live view)

set_smart_input_as(arg1, context)

set_smart_input_text(socket, text \\ "\n")

sigil_p(route, extra)
(macro)

smart_input_name(component)

toggle_expanded(target, js \\ %JS{})

toggle_expanded(target, btn, class, js \\ %JS{})

toggle_minimize(target, status, js \\ %JS{})