Bonfire.UI.Common.SmartInputModule behaviour (bonfire_umbrella v0.9.8-cooperation-beta.30)

Find a smart input module via the object type(s) it can create, or vice versa. Backed by a global cache of known smart input modules.

To add a module to this list, you should declare @behaviour Bonfire.UI.Common.SmartInputModule in it and define a smart_input_module/0 function which returns a list of object types that it can create.

Example:

@behaviour Bonfire.UI.Common.SmartInputModule
def smart_input_module, do: [:page, Bonfire.Pages.Page]

You can then open the smart input composer / object creator using that declared type, for example:

<Bonfire.UI.Common.SmartInputButtonLive
  create_object_type={:page}
  prompt={l("New page")}
  icon="mdi:pencil"
/>

Callbacks

smart_input_module()

@callback smart_input_module() :: any()

Declares a smart input module

Functions

app_modules()

modules()

@spec modules() :: [atom()]

Callback implementation for Bonfire.Common.ExtensionBehaviour.modules/0.

smart_input_modules_types()

Returns a list of smart input modules and the object type(s) it can create, and vice versa.