EctoSparkles.DataMigration.Config (bonfire_umbrella v0.9.8-cooperation-beta.18)

Types

t()

@type t() :: %EctoSparkles.DataMigration.Config{
  async: boolean(),
  batch_size: integer(),
  first_id: any(),
  repo: atom(),
  throttle_ms: integer()
}

Configuration for a DataMigration behaviour module, used by the DataMigration.Runner.

batch size: how many elements from your table to migrate at a time.

throttle time: the amount of downtime the runner should sleep between batches.

async: Whether to run the migration in an async process, meaning the execution of the rest of the migrations (and the app startup if you're auto-migrating on start) won't be delayed. WARNING: this means the migration will be marked as done as soon as it starts, so if the process is interrupted it won't be re-run automatically.

first_id: The very first ID when sorting UUIDs in ascending order. If you use integer IDs instead, this would be 0.