mix bonfire.extension.copy_migrations (bonfire_umbrella v0.9.8-cooperation-beta.30)

Functions

copy(extension_paths, dest_path, opts)

maybe_copy(extensions \\ nil, opts)

run(args)

Usage: just mix bonfire.extension.copy_migrations my_extension or just mix bonfire.extension.copy_migrations

NOTE: if you don't specify what extension(s) to include, it will automatically include all extensions which:

start with bonfire_

and are included in the top-level app (not dependencies of dependencies)

Optional args:

--force (to not ask for confirmation before copying, or to overwrite existing migration files) --from priv/repo/migrations (to change the source repo paths, relative to each extension path) --to priv/repo/migrations (to change the target repo path (defaults to current flavour's migrations) relative to working directory) --repo MyRepo (to specify what repo to migrate after)