Needle.ULID (bonfire_umbrella v0.9.8-cooperation-beta.26)

An Ecto type for ULID strings.

Functions

bingenerate(timestamp \\ System.system_time(:millisecond))

Generates a binary ULID.

If a value is provided for timestamp, the generated ULID will be for the provided timestamp. Otherwise, a ULID will be generated for the current time.

Arguments:

timestamp: A Unix timestamp with millisecond precision.

bintimestamp(arg)

cast(value)

Casts a 26-byte encoded string to ULID or a 16-byte binary unchanged

cast!(value)

Same as cast/1 but raises Ecto.CastError on invalid arguments.

dump(encoded)

Converts a Crockford Base32 encoded ULID into a binary.

dump!(encoded)

embed_as(_)

Callback implementation for Ecto.Type.embed_as/1.

encoded_randomness(arg)

Returns the randomness portion of the encoded ulid

encoded_timestamp(arg)

Returns the timestamp portion of the encoded ulid

equal?(term1, term2)

Callback implementation for Ecto.Type.equal?/2.

generate(timestamp \\ System.system_time(:millisecond))

Generates a Crockford Base32 encoded ULID.

If a value is provided for timestamp, the generated ULID will be for the provided timestamp. Otherwise, a ULID will be generated for the current time.

Arguments:

timestamp: A Unix timestamp with millisecond precision.

load(bytes)

Converts a binary ULID into a Crockford Base32 encoded string.

synthesise!(x)

translates alphanumerics into a sentinel ulid value

timestamp(encoded)

Returns the timestamp of an encoded or unencoded ULID

type()

The underlying schema type.