Paginator.PageInfo (bonfire_umbrella v0.9.8-cooperation-beta.28)

Defines page page_info.

Fields

start_cursor - an opaque cursor representing the first row of the current page, to be used with the before query parameter.

end_cursor - an opaque cursor representing the last row of the current page, to be used with the after query parameter.

limit - the maximum number of edges that can be contained in this page.

page_count - the number of edges on the current page.

total_count - the total number of edges matching the query.

total_count_cap_exceeded - a boolean indicating whether the :total_count_limit was exceeded.

Types

opaque_cursor()

@type opaque_cursor() :: String.t()

t()

@type t() :: %Paginator.PageInfo{
  end_cursor: opaque_cursor() | nil,
  limit: integer(),
  page_count: integer() | nil,
  start_cursor: opaque_cursor() | nil,
  total_count: integer() | nil,
  total_count_cap_exceeded: boolean() | nil
}