PaperTrail (bonfire_umbrella v0.9.8-cooperation-beta.26)

Functions

add_prefix(schema, prefix)

See PaperTrail.Serializer.add_prefix/2.

delete(model_or_changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec delete(
  model_or_changeset :: model | Ecto.Changeset.t(model),
  options :: Keyword.t()
) ::
  {:ok, %{model: model, version: PaperTrail.Version.t()}}
  | {:error, Ecto.Changeset.t(model) | term()}
when model: struct()

Deletes a record from the database with a related version insertion in one transaction

delete!(model_or_changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec delete!(
  model_or_changeset :: model | Ecto.Changeset.t(model),
  options :: Keyword.t()
) :: model
when model: struct()

Same as delete/2 but returns only the model struct or raises if the changeset is invalid.

get_current_model(version)

See PaperTrail.VersionQueries.get_current_model/1.

get_item_type(data)

See PaperTrail.Serializer.get_item_type/1.

get_model_id(model)

See PaperTrail.Serializer.get_model_id/1.

get_sequence_id(table_name)

See PaperTrail.Serializer.get_sequence_id/1.

get_version(record)

See PaperTrail.VersionQueries.get_version/1.

get_version(model_or_record, id_or_options)

See PaperTrail.VersionQueries.get_version/2.

get_version(model, id, options)

See PaperTrail.VersionQueries.get_version/3.

get_versions(record)

See PaperTrail.VersionQueries.get_versions/1.

get_versions(model_or_record, id_or_options)

See PaperTrail.VersionQueries.get_versions/2.

get_versions(model, id, options)

See PaperTrail.VersionQueries.get_versions/3.

has_version?(record)

See PaperTrail.VersionQueries.has_version?/1.

has_version?(model_or_record, id_or_options)

See PaperTrail.VersionQueries.has_version?/2.

has_version?(model, id, options)

See PaperTrail.VersionQueries.has_version?/3.

initialise(model, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, version_key: :version])

Explicitly inserts a non-versioned already existing record into the Versions table

insert(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec insert(
  changeset :: Ecto.Changeset.t(model),
  options :: Keyword.t()
) ::
  {:ok, %{model: model, version: PaperTrail.Version.t()}}
  | {:error, Ecto.Changeset.t(model) | term()}
when model: struct()

Inserts a record to the database with a related version insertion in one transaction

insert!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec insert!(changeset :: Ecto.Changeset.t(model), options :: Keyword.t()) :: model
when model: struct()

Same as insert/2 but returns only the model struct or raises if the changeset is invalid.

insert_or_update(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec insert_or_update(changeset :: Ecto.Changeset.t(model), options :: Keyword.t()) ::
  {:ok, %{model: model, version: PaperTrail.Version.t()}}
  | {:error, Ecto.Changeset.t(model) | term()}
when model: struct()

Upserts a record to the database with a related version insertion in one transaction.

insert_or_update!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec insert_or_update!(changeset :: Ecto.Changeset.t(model), options :: Keyword.t()) ::
  model
when model: struct()

Same as insert_or_update/2 but returns only the model struct or raises if the changeset is invalid.

make_version_struct(version, model, options)

See PaperTrail.Serializer.make_version_struct/3.

serialize(data)

See PaperTrail.Serializer.serialize/1.

update(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec update(changeset :: Ecto.Changeset.t(model), options :: Keyword.t()) ::
  {:ok, %{model: model, version: PaperTrail.Version.t()}}
  | {:error, Ecto.Changeset.t(model) | term()}
when model: struct()

Updates a record from the database with a related version insertion in one transaction

update!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []])

@spec update!(changeset :: Ecto.Changeset.t(model), options :: Keyword.t()) :: model
when model: struct()

Same as update/2 but returns only the model struct or raises if the changeset is invalid.