Untangle (bonfire_umbrella v0.9.8-cooperation-beta.30)

Logging and debug printing that include location information

Functions

custom_dbg(code, options, env)

Custom backend for Kernel.dbg/2. This function provides a backend for Kernel.dbg/2. This function:

may log or print information about the given env

may log or print information about code and its returned value (using opts to inspect terms)

returns the value returned by evaluating code

debug(thing, label \\ "")
(macro)

Like dump, but for logging at debug level

dump(thing, label \\ "")
(macro)

IO.inspect but outputs to Logger with position information, an optional label and configured not to truncate output too much.

error(thing, label \\ "")
(macro)

Similar to dump, but for logging at error level, and returns an error tuple:

an error tuple with the label, if any

an error tuple with the passed value otherwise

iex> error(:value)

[error] :value

iex> error({:error, :value})

[error] :value

iex> error(:value, "with label")

[error] with label: :value

iex> error({:error, :value}, "with label")

[error] with label: :value

format_stacktrace_sliced(stacktrace, starts \\ 1, ends \\ 5)

info(thing, label \\ "")
(macro)

Like dump, but for logging at info level

log_level?(level)

maybe_dbg(thing, label \\ "", options)
(macro)

Like debug, but will do nothing unless the :debug option is truthy

maybe_info(thing, label \\ "", options)
(macro)

Like maybe_dbg, but requires the :verbose option to be set. Intended for large outputs.

smart(thing, label \\ "", options)
(macro)

Tries to 'do what i mean'. Requires the debug option to be set regardless. If verbose is also set, will inspect else will attempt to print some (hopefully smaller) type-dependent summary of the data (list length, map keys).

warn(thing, label \\ "")
(macro)

Like dump, but for logging at warn level