Zest (bonfire_umbrella v0.4.0-classic-beta.62)

Types

intercept_type()

@type intercept_type() :: :rescue | :catch | :exit

interceptor()

@type interceptor() ::
  (intercept_type(), error :: term(), maybe_stacktrace :: term() -> term())

Functions

each(arg1, fun)

Iterates over a collections, calling the provided effectful function with each item.

each(a, b, fun)

Iterates over two collections, calling the provided effectful function with each pair of items

hijack(fun, jack)

@spec hijack(function(), jack :: (function(), [term()] -> term())) :: function()

You take on the role of the apply function in this exciting function that wraps execution of a function such that your function is responsible for calling it.

in_scope(scopes, fun)

@spec in_scope(Keyword.t(), function()) :: term()

Add some debug information to the context for the duration of a function's execution. If a raise, throw or exit occurs, the context will be pretty printed to the screen to aid with debugging.

intercept(fun, interceptor)

@spec intercept(function(), interceptor()) :: function()

Catches errors and exceptions, invoking an interceptor function

rethrow(atom, e, stacktrace)

@spec rethrow(intercept_type(), error :: term(), maybe_stacktrace :: term()) :: none()

An interceptor function which simply rethrows/reraises/re-exits

rethrowing(fun)

@spec rethrowing(function()) :: function()

Wraps an interceptor or nullary function into an interceptor function such that after the execution of the provided function, the error or exception will be rethrown.

scope(list, expr)
(macro)

Add some debug information to the context for the duration of a block or expression. If a raise, throw or exit occurs, the context will be pretty printed to the screen to aid with debugging.

Examples:

scope [foo: :bar[, assert(true == false)

scope [foo: :bar] do assert true == false end

scoped(scopes, fun)

@spec scoped(Keyword.t(), region :: function()) :: function()

Wrap a function such that it is as if its body was wrapped in scope/2