ActivityPub.Actor (bonfire_umbrella v0.9.8-cooperation-beta.15)

An ActivityPub Actor type and functions for dealing with actors.

Types

t()

@type t() :: %ActivityPub.Actor{
  ap_id: term(),
  data: term(),
  deactivated: term(),
  id: term(),
  keys: term(),
  local: term(),
  pointer: term(),
  pointer_id: term(),
  updated_at: term(),
  username: term()
}

Functions

actor?(arg1)

actor_url(username)

check_actor_is_active(actor)

deactivate(actor)

delete(actor)

fetch_by_username(username, opts \\ [])

Fetches a remote actor by username in username@domain.tld format

format_remote_actor(object)

format_username(ap_id)

format_username(ap_id, nick)

get_cached(id)

get_cached!(opts)

get_cached_or_fetch(username_or_uri)

Tries to get a local actor by username or tries to fetch it remotely if username is provided in username@domain.tld format.

get_external_followers(actor)

get_followers(actor)

get_followings(actor)

invalidate_cache(actor)

maybe_create_actor_from_object(actor)

reactivate(actor)

save_actor_tombstone(actor, tombstone)

set_cache(actor)

swap_or_create_actor_tombstone(actor)

update_actor(actor_id)

@spec update_actor(String.t()) :: {:ok, t()} | {:error, any()}

Updates an existing actor struct by its AP ID.

update_actor(actor_id, data)