ActivityPub.Object (bonfire_umbrella v0.9.8-cooperation-beta.26)

Types

t()

@type t() :: %ActivityPub.Object{
  __meta__: term(),
  activity: term(),
  data: term(),
  id: term(),
  inserted_at: term(),
  is_object: term(),
  local: term(),
  object: term(),
  pointer: term(),
  pointer_id: term(),
  public: term(),
  updated_at: term()
}

Functions

actor_from_data(actor)

actor_id_from_data(id)

changeset(attrs)

changeset(object, attrs)

common_changeset(object)

delete(object)

do_insert(attrs)

do_update_existing(object, attrs)

fetch_latest_block(map1, map2)

fetch_latest_follow(follower_id, followed_id)

get_activity_for_object_ap_id(ap_id, verb \\ "Create")

get_ap_id(_)

get_cached(id)

get_cached!(opts)

get_existing_announce(actor, map)

Retruns an existing announce activity if the notice has already been announced

get_existing_like(actor, object_id)

Returns an existing like if a user already liked an object

get_follow_activity(follow_object, followed)

get_outbox_for_actor(ap_id, page \\ 1)

get_outbox_for_instance()

get_uncached(opts)

hard_delete(object)

hashtags(arg1)

invalidate_cache(object)

make_tombstone(map, deleted \\ DateTime.utc_now())

maybe_fetch(ap_id, arg2)

maybe_upsert(arg1, existing_object, attrs)

normalize(object, fetch_remote? \\ true, pointer \\ nil)

normalize_actors(params)

normalize_params(params, activity_id \\ nil, pointer \\ nil)

object_url(id)

prepare_data(data, local \\ false, pointer \\ nil, associated_activity \\ nil)

Prepares a struct to be inserted into the objects table

self_replies_ids(object, limit)

set_cache(object)

swap_object_with_tombstone(object)

update_and_set_cache(changeset)

update_changeset(object, attrs)

update_existing(object_id, attrs)

update_state(activity, type, state)

Updates a follow activity's state (for locked accounts).