Bonfire.Boundaries (bonfire_umbrella v0.9.8-cooperation-beta.21)

Functions

acls_from_preset_boundary_names(presets)

boundaries_normalise(text)

boundaries_or_default(to_boundaries, context \\ [])

can?(circle_name, verbs, object, opts \\ [])

default_boundaries(context \\ [])

find_caretaker_stereotype(caretaker, stereotype, from)

find_caretaker_stereotypes(caretaker, stereotypes, from \\ Pointer)

find_caretaker_stereotypes_q(caretaker, stereotypes, from)

list_grants_on(things)

list_grants_on(things, verbs)

eg: list_grants_on(id, [:see, :read])

list_object_acls(object, opts \\ [])

list_object_boundaries(object, opts \\ [])

load_pointer(item, opts)

load_pointers(items, opts)

Loads binaries according to boundaries (which are assumed to be ULID pointer IDs). Lists which are iterated and return a [sub]list with only permitted pointers.

pointer_permitted?(item, opts)

preset_boundary_from_acl(acl, object_type \\ nil)

preset_boundary_tuple_from_acl(acl, object_type \\ nil)

preset_name(boundaries, include_remote? \\ false)

replace_boundaries(creator, object, previous_preset, opts)

set_boundaries(creator, object, opts)

take_care_of!(things, user)

Assigns the user as the caretaker of the given object or objects, replacing the existing caretaker, if any.

user_default_boundaries()

users_grants_on(users, things)

users_grants_on(users, things, verbs)