Bonfire.Boundaries.Grants (bonfire_umbrella v0.9.0-classic-beta.7)

a grant applies to a subject

Functions

changeset(grant \\ %Grant{}, attrs)

changeset(atom, attrs, opts)

create(attrs)

create(attrs, opts)

grant(subject_id, acl_id, verb, value, opts \\ [])

Grant takes three parameters:

subject_id: who we are granting access to

acl_id: what ACL we're applying a grant to

verb: which verb/action

value: true, false, or nil

grant_role(subject_id, acl_id, role, opts \\ [])

grants()

list(opts)

Lists the grants permitted to see.

list_my(user)

Lists the grants we are the registered caretakers of that we are permitted to see. If any are created without permitting the user to see them, they will not be shown.

list_my_q(user)

query for list_my

list_q(opts)

remove_subject_from_acl(subject, acls)

upsert_or_delete(attrs, opts)