Bonfire.Boundaries.Grants (bonfire_umbrella v0.9.8-cooperation-beta.5)

A grant is part of an Acl, and defines a permission (value boolean on a verb) for a subject

Functions

changeset(grant \\ %Grant{}, attrs)

changeset(atom, attrs, opts)

create(attrs)

create(attrs, opts)

get(slug)

grant(subject_id, acl_id, verb, value, opts \\ [])

Edits or adds a grant to an Acl

Takes three parameters:

subject_id: who we are granting access to

acl_id: what ACL we're applying a grant to

verb: which verb/action

value: true, false, or nil

grant_role(subject_id, acl_id, role, opts \\ [])

Edits or adds grants to an Acl based on a role

grants()

grants_to_tuples(creator, grants)

list(opts)

Lists the grants permitted to see.

list_for_acl(acl, opts)

list_my(user)

Lists the grants we are the registered caretakers of that we are permitted to see. If any are created without permitting the user to see them, they will not be shown.

list_q(opts)

remove_subject_from_acl(subject, acls)

subject_grants(grants)

subject_verb_grants(grants)

subjects(grants)

upsert_or_delete(attrs, opts)

verb_subject_grant(grants)