Bonfire.Common.Config (bonfire_umbrella v0.4.0-classic-beta.28)

Functions

compilation_error(error)
(macro)

delete(key, otp_app \\ top_level_otp_app())

get!(key, otp_app \\ top_level_otp_app())

get(key_or_keys, default \\ nil, otp_app \\ top_level_otp_app())

Get config value for a config key (optionally from a specific OTP app or Bonfire extension)

get_ext!(module_or_otp_app)

get_ext!(module_or_otp_app, key)

get_ext(module_or_otp_app)

Get all config keys/values for a specific Bonfire extension or OTP app

get_ext(module_or_otp_app, key, default \\ nil)

Get config value for a Bonfire extension or OTP app config key

keys_tree(keys)

Constructs a key path for settings/config, which always starts with an app or extension name (which defaults to the main OTP app)

iex> keys_tree([:bonfire_me, Bonfire.Me.Users]) [:bonfire_me, Bonfire.Me.Users]

iex> keys_tree(Bonfire.Me.Users) [:bonfire_me, Bonfire.Me.Users]

iex> keys_tree(:bonfire_me) [:bonfire_me]

iex> keys_tree(:random_atom) [:bonfire, :random_atom]

iex>keys_tree([:random_atom, :sub_key]) [:bonfire, :random_atom, :sub_key]

module_enabled?(module)

See Bonfire.Common.Extend.module_enabled?/1.

put(tree)

put(key, value, otp_app \\ nil)

repo()

require_extension_config!(extension)

Stop if an Elixir module or extension / OTP app doesn't have configuration keys set up

top_level_otp_app()