Bonfire.Common.Utils (bonfire_umbrella v0.9.8-cooperation-beta.29)

Various very commonly used utility functions for the Bonfire application.

Functions

apply_error(error, args, opts)

apply_task(function \\ :async, fun, opts \\ [])

Runs a function asynchronously in a Task. Simply a shorthand for calling functions in Task and Task.Supervisor but with support for multi-tenancy in the spawned process.

Task.async/1 the caller creates a new process links and monitors it. Once the task action finishes, a message is sent to the caller with the result. Task.await/2 is used to read the message sent by the task. When using async, you must await a reply as they are always sent.

Task.start_link/1 is suggested instead if you are not expecting a reply. It starts a statically supervised task as part of a supervision tree, linked to the calling process (meaning it will be stopped when the caller stops).

Task.start/1 can be used for fire-and-forget tasks, like side-effects, when you have no interest on its results nor if it completes successfully (because if the server is shut down it won't be restarted).

For more serious tasks, consider using Oban or apply_task_supervised/3 for supervised tasks when possible:

Task.Supervisor.start_child/2 allows you to start a fire-and-forget task when you don't care about its results or if it completes successfully or not.

Task.Supervisor.async/2 + Task.await/2 allows you to execute tasks concurrently and retrieve its result. If the task fails, the caller will also fail.

apply_task_supervised(function \\ :async, fun, opts \\ [])

current_account(current_account_or_socket_or_opts, recursing \\ false)

Returns the current account from socket, assigns, or options.

current_account_and_or_user_ids(assigns)

current_account_auth!(context, password)

current_account_id(current_account_or_socket_or_opts, recursing \\ false)

current_user(current_user_or_socket_or_opts, recursing \\ false)

Returns the current user from socket, assigns, or options.

current_user_auth!(context, password)

current_user_id(current_user_or_socket_or_opts, recursing \\ false)

Returns the current user ID from socket, assigns, or options.

current_user_required!(context)

e(val, fallback)

Returns a value, or a fallback if nil/false

e(object, key, fallback)

Extracts a value from a map (and various other data structures), or returns a fallback if not present or empty. If more arguments are provided it looks for nested data (with the last argument always being the fallback).

e(object, key1, key2, fallback)

Returns a value from a nested map, or a fallback if not present

e(object, key1, key2, key3, fallback)

e(object, key1, key2, key3, key4, fallback)

e(object, key1, key2, key3, key4, key5, fallback)

e(object, key1, key2, key3, key4, key5, key6, fallback)

empty?(v)

Returns true if the given value is nil, an empty map, an empty list, or an empty string.

maybe(val, change_fn)

Applies change_fn if the first parameter is not nil.

maybe_apply(module, fun, args \\ [], opts \\ [])

Helpers for calling hypothetical functions in other modules. Returns the result of calling a function with the given arguments, or the result of fallback function if the primary function is not defined (by default just logging an error message).

maybe_from_opts(opts, key, fallback \\ nil)

Returns the value of a key from options keyword list or map, or a fallback if not present or empty.

ok_unwrap(val, fallback \\ nil)

Unwraps an {:ok, val} tuple, returning the value, or returns a fallback value (nil by default) if the tuple is {:error, _} or :error.

round_nearest(num)

round_nearest(num, target)

socket_connected?(socket)

to_options(user_or_socket_or_opts)

Converts a map, user, socket, tuple, etc, to a keyword list for standardised use as function options.