Bonfire.Social.FeedActivities (bonfire_umbrella v0.9.8-cooperation-beta.24)

Functions

cast(changeset, feed_ids)

cast(changeset, creator, opts)

count(filters \\ [], opts \\ [])

count_subjects(filters \\ [], opts \\ [])

count_total()

delete(filters)

Remove activities from feeds, using specific filters

delete(objects, by_field)

Remove one or more activities from all feeds

feed(feed, opts \\ [])

Gets a feed by id or ids or a thing/things containing an id/ids.

feed_contains?(feed_name, object, opts \\ [])

feed_ids_and_opts(feed_name, opts)

feed_many_paginated(query, opts)

feed_name(name, current_user_or_socket)

feed_paginated(filters \\ [], opts \\ [])

Return a page of Feed Activities (reverse chronological) + pagination metadata TODO: consolidate with feed/2

feed_paginated(filters, opts, query)

feed_with_object(feed_name, object, opts \\ [])

feeds_for_activity(id)

Gets a list of feed ids this activity was published to from the database.

Currently only used by the ActivityPub integration.

filters_from_opts(opts)

get_feed_ids(options)

Computes the feed ids for get_feed_publishes/2.

Options:

:inbox - list of objects whose inbox we should attempt to insert into.

:outbox - list of objects whose outbox we should attempt to insert into.

:notifications - list of objects whose notifications we should attempt to insert into.

:feeds - list of ids (or objects containing IDs of feeds to post to.

get_feed_publishes(options)

Creates the underlying data for put_feed_publishes/2.

mark_all_seen(feed_id, opts)

maybe_feed_publish(subject, verb_or_activity, object, feeds, opts)

Creates a new local activity or takes an existing one and publishes to specified feeds

my_feed(opts, home_feed_ids \\ nil)

Gets a user's home feed, a combination of all feeds the user is subscribed to.

publish(subject, verb_or_activity, object, opts \\ [])

Creates a new local activity and publishes to appropriate feeds TODO: make this re-use the changeset-based code like in Epics instead of duplicating logic (currently it is only used in VF extension anyway)

put_feed_publishes(changeset, options)

Arranges for an insert changeset to also publish to feeds related to some objects.

Options: see get_feed_ids/1

query(filters \\ [], opts \\ [], query \\ default_query())

Return a boundarised query for a feed

schema_module()

Callback implementation for Bonfire.Common.ContextModule.schema_module/0.

skip_verbs_default()

the_object(object)

to_feed_options(socket_or_opts)

unseen_count(feed_id, opts)