Bonfire.UI.Common (bonfire_umbrella v0.9.8-cooperation-beta.26)

A library of common utils and helpers used across Bonfire extensions

Functions

assign_error(socket, msg, pid \\ self())

assign_flash(socket_or_conn, type, message, assigns \\ %{}, pid \\ self())

assign_generic(socket_or_conn, assigns)

assign_generic(socket, key, value)

assign_generic_global(socket, assigns)

assign_global(socket, assigns)

assign_global(socket, key, value)

assigns_clean(tuple)

assigns_merge(assigns, new)

assigns_merge(socket, assigns, new)

assigns_minimal(assigns)

assigns_subscribe(socket, assign_names)

batch_update_many_async(assigns_sockets, many_opts, opts)

batch_update_many_async(current_user, assigns_sockets, many_opts, opts)

can?(subject, verbs, object, opts \\ [])

cast_public(socket, assigns_to_broadcast)

Warning: this will set assigns for any/all users who subscribe to them. You want to cast_self/2 instead if dealing with user-specific actions or private data.

cast_self(socket, assigns_to_broadcast)

copy_go(arg1)

copies the go param into a query string, if any

current_url(socket_or_assigns, default \\ nil)

current_user_or_remote_interaction(socket, verb, object)

go_query(url)

Generate a query string adding a go redirection path to the URI (for redirecting somewhere after auth flows). It is recommended to use set_go_after/2 where possible instead.

hero_icons_list()

live_aliases(aliases, path, live_view, action \\ nil, opts \\ [])
(macro)

live_render_component(conn, load_live_component)

Special LiveView helper function which allows loading LiveComponents in regular Phoenix views: live_render_component(@conn, MyLiveComponent)

live_render_with_conn(conn, live_view)

live_upload_files(current_user, metadata, socket)

markdown(content)

maybe_apply_or_ret(assigns, mod, fun)

maybe_assign(socket, key, value)

maybe_component(module, context \\ [])

maybe_consume_uploaded_entries(socket, key, fun)

maybe_consume_uploaded_entry(socket, key, fun)

maybe_cute_gif()

maybe_last_sentry_event_id()

maybe_push_event(socket, name, data)

maybe_send_update(component, id, assigns, opts \\ [])

maybe_stream_insert(socket, name, items, opts)

Inserts one or many items in an existing stream. See Phoenix.LiveView.stream_insert/4 for opts.

opts_for_update_many_async(arg, opts)

patch_to(socket_or_conn, to \\ nil, opts \\ [])

path_fallback(socket_or_conn, opts)

redirect_to(socket_or_conn, to \\ nil, opts \\ [])

redirect_to_previous_go(conn, params, default, current_path)

render_sface_or_native(opts \\ [])
(macro)

rich(content, opts \\ [])

self_subscribe(socket, assign_names)

Subscribe to assigns targeted at the current account/user

send_self(socket \\ nil, assigns)

send_self_global(socket \\ nil, assigns)

set_go_after(conn, path \\ nil)

Save a go redirection path in the session (for redirecting somewhere after auth flows)

templated(content, data \\ %{})

templated_or_remote_markdown(content, data \\ nil)

the_object(assigns)

undead(socket, fun, return_key \\ :noreply)

undead_mount(socket, fun)

Run a function and expects tuple. If anything else is returned, like an error, a flash message is shown to the user.

undead_on_mount(socket, fun)

undead_params(socket, fun)

update_many_async(assigns_sockets, opts)

update_many_async(current_user, assigns_sockets, opts)